+++ sponge-8/src/main/java/dev/qixils/crowdcontrol/plugin/sponge8/utils/BlockFinder.java
@@ -1,14 +1,13 @@
 package dev.qixils.crowdcontrol.plugin.sponge8.utils;
 
 import dev.qixils.crowdcontrol.common.util.AbstractBlockFinder;
-import dev.qixils.crowdcontrol.common.util.CommonTags;
 import org.jetbrains.annotations.NotNull;
 import org.spongepowered.api.block.BlockState;
-import org.spongepowered.api.block.BlockType;
 import org.spongepowered.api.data.Key;
 import org.spongepowered.api.data.Keys;
+import org.spongepowered.api.data.type.MatterType;
+import org.spongepowered.api.data.type.MatterTypes;
 import org.spongepowered.api.data.value.Value;
-import org.spongepowered.api.registry.RegistryTypes;
 import org.spongepowered.api.world.server.ServerLocation;
 import org.spongepowered.api.world.server.ServerWorld;
 import org.spongepowered.math.vector.Vector3i;
@@ -19,7 +18,9 @@ import java.util.function.Predicate;
 public final class BlockFinder extends AbstractBlockFinder<ServerLocation, Vector3i, ServerWorld> {
 	public static final Predicate<ServerLocation> SPAWNING_SPACE = location ->
 			isPassable(location.block())
+					&& !isLiquid(location.block())
 					&& isPassable(location.add(0, 1, 0).block())
+				    && !isLiquid(location.add(0, 1, 0).block())
 					&& isSolid(location.sub(0, 1, 0).block());
 
 	private BlockFinder(ServerWorld origin, List<Vector3i> positions, Predicate<ServerLocation> locationValidator) {
@@ -28,6 +29,18 @@ public final class BlockFinder extends AbstractBlockFinder<ServerLocation, Vecto
 
 	public static boolean isProperty(BlockState block, Key<Value<Boolean>> propertyKey, boolean def) {
 		return block.get(propertyKey).orElse(def);
+	}
+
+	public static boolean isMatter(BlockState block, MatterType matter) {
+		return block.get(Keys.MATTER_TYPE).map(actual -> actual.equals(matter)).orElse(false);
+	}
+
+	public static boolean isLiquid(BlockState block) {
+		return isMatter(block, MatterTypes.LIQUID.get());
+	}
+
+	public static boolean isAir(BlockState block) {
+		return isMatter(block, MatterTypes.GAS.get());
 	}
 
 	public static boolean isPassable(BlockState block) {
@@ -40,14 +53,6 @@ public final class BlockFinder extends AbstractBlockFinder<ServerLocation, Vecto
 
 	public static boolean isReplaceable(BlockState block) {
 		return isProperty(block, Keys.IS_REPLACEABLE, false);
-	}
-
-	public static boolean isAir(BlockState block) {
-		return isAir(block.type());
-	}
-
-	public static boolean isAir(BlockType block) {
-		return CommonTags.AIR.contains(block.key(RegistryTypes.BLOCK_TYPE));
 	}
 
 	public static BlockFinderBuilder builder() {