+++ CHANGELOG.md
@@ -12,7 +12,7 @@ A list of all changes made to the software in reverse chronological order.
 - Added extra diagnostic information to `/crowdcontrol status`
 - Changed the Bucket Clutch effect to work in the nether by using a cobweb instead of water
 - Changed the Bucket Clutch effect to teleport the player even higher where possible
-- Changed the Chorus Fruit effect to better match vanilla behavior on Fabric and Paper
+- Changed the Chorus Fruit effect to better match vanilla behavior
 - Fixed warnings about duplicate effects on some modpacks
 - Fixed summon effects on Sponge 10+ (1.19+)
 - Fixed killer rabbits spawning more than intentional on Paper servers